% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class Example.R
\name{class}
\alias{class}
\title{SVRS}
\usage{
class(f, N, low, high, maxf)
}
\arguments{
\item{f}{pdf sampled from}

\item{N}{number of attempted samples}

\item{low}{lower bound of the support of f}

\item{high}{upper}

\item{maxf}{bound of f}
}
\value{
vector of samples from pdf
}
\description{
Description
}
\examples{
betaPDF <- function(x) {ifelse(0<x & x<1, 2*x, 0)}
class(betaPDF, 100, 0, 1, 2)
}
