% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectSample.R
\name{oneDSAMPLE}
\alias{oneDSAMPLE}
\title{Single variable rejection sampling}
\usage{
oneDSAMPLE(f, N, lb, ub, maxf)
}
\arguments{
\item{f}{the pdf that we are sampling from}

\item{N}{the number of attempted samples}

\item{lb}{the lower bound of support f}

\item{nb}{the upper bound of support f}

\item{mxaf}{bound on f}
}
\value{
A vector containing samples from pdf
}
\description{
This function implements single variable rejection samping for rvs
 with bounded support and which have have bounded pdf. we expect N/maxf sample
}
\details{
Here are more details about the algorithm that we are using
}
\examples{

betaPDF <- function(x){
 ifelse(0<x&x <1, 2*x, 0)}
oneDSAMPLE(f = betaPDF, N=100, lb=0, ub = 1, maxf = 2)
hist(oneDSAMPLE(f = betaPDF, N=100, lb=0, ub = 1, maxf = 2))
}
