% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single Sample.R
\name{svrs}
\alias{svrs}
\title{Single Variable Random Samples}
\usage{
svrs(PDF, N, a, b)
}
\arguments{
\item{PDF}{a function to be sampled from}

\item{N}{number of attempted samples}

\item{a}{lower bound of the support of PDF}

\item{b}{upper bound of the support of PDF}
}
\value{
a dataframe of samples from PDF
}
\description{
Description
}
\examples{
betaPDF <- function(x) {ifelse(0<x & x<1, 2*x, 0)}
svrs(betaPDF, 100, 0, 1)

unifPDF <- function(x) {ifelse(0<x & x<1, 1, 0)}
svrs(betaPDF, 100, 0, 1)
}
